Split WebAuthn request types
