+++ Sources/Panels/BrowserWebAuthnBridgeMessageKind.swift
@@ -1,0 +1,5 @@
+enum BrowserWebAuthnBridgeMessageKind: String {
+    case capabilities
+    case createCredential
+    case getCredential
+}
