+++ Sources/Panels/BrowserWebAuthnMessageEnvelope.swift
@@ -1,0 +1,4 @@
+struct BrowserWebAuthnMessageEnvelope {
+    let kind: BrowserWebAuthnBridgeMessageKind
+    let payloadJSON: String?
+}
