+++ Sources/Panels/BrowserWebAuthnTransport.swift
@@ -1,0 +1,7 @@
+enum BrowserWebAuthnTransport: String {
+    case ble
+    case hybrid
+    case `internal`
+    case nfc
+    case usb
+}
