+++ Sources/Panels/BrowserWebAuthnTransportSummary.swift
@@ -1,0 +1,63 @@
+struct BrowserWebAuthnTransportSummary {
+    let containsBluetooth: Bool
+    let containsHybrid: Bool
+    let containsInternal: Bool
+    let containsSecurityKeyTransport: Bool
+    let containsUnspecifiedTransport: Bool
+
+    init(descriptors: [BrowserWebAuthnCredentialDescriptor]) {
+        var containsBluetooth = false
+        var containsHybrid = false
+        var containsInternal = false
+        var containsSecurityKeyTransport = false
+        var containsUnspecifiedTransport = false
+
+        for descriptor in descriptors where descriptor.isPublicKeyCredential {
+            let transports = descriptor.normalizedTransports
+            if transports.isEmpty {
+                containsUnspecifiedTransport = true
+                continue
+            }
+
+            for transport in transports {
+                switch transport {
+                case .ble:
+                    containsBluetooth = true
+                    containsSecurityKeyTransport = true
+                case .hybrid:
+                    containsHybrid = true
+                case .internal:
+                    containsInternal = true
+                case .nfc, .usb:
+                    containsSecurityKeyTransport = true
+                }
+            }
+        }
+
+        self.containsBluetooth = containsBluetooth
+        self.containsHybrid = containsHybrid
+        self.containsInternal = containsInternal
+        self.containsSecurityKeyTransport = containsSecurityKeyTransport
+        self.containsUnspecifiedTransport = containsUnspecifiedTransport
+    }
+
+    var allowsPlatformCredentials: Bool {
+        containsInternal || containsHybrid || containsUnspecifiedTransport
+    }
+
+    var allowsSecurityKeyCredentials: Bool {
+        containsSecurityKeyTransport || containsHybrid || containsUnspecifiedTransport
+    }
+
+    var needsBluetoothPreparation: Bool {
+        containsBluetooth || containsHybrid
+    }
+
+    var shouldShowHybridTransport: Bool {
+        containsHybrid || containsUnspecifiedTransport
+    }
+
+    var prefersSecurityKeysFirst: Bool {
+        containsSecurityKeyTransport && !containsInternal && !containsHybrid && !containsUnspecifiedTransport
+    }
+}
